ref(namespace): Move submit too Submission
Add HandlerResponseInterface
Communication with submissionBundle should live in the Submission
namespace
