+++ Submission/HandleResponseInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EMS\FormBundle\Submission;
+
+interface HandleResponseInterface
+{
+    public function getStatus(): string;
+    public function getResponse(): string;
+}
